pyproject.toml
[project]
name = "live-audio-translator"
version = "0.1.0"
description = "Real-time audio translation between English and French"
readme = "PROJECT_DEFINITION.md"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper",
    "argostranslate",
    "PyQt6",
    "pyaudio",
    "numpy",
]

[project.optional-dependencies]
dev = [
    "ruff",
]

[project.scripts]
live-translator = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
